issue-6385: extracted block/byte layer visitors into layer.h/cpp, documented this thing, documented fresh byte layer-related things
