+++ cloud/filestore/libs/storage/tablet/model/blob_builder.h
@@ -3,6 +3,7 @@
 #include "public.h"
 
 #include "blob.h"
+#include "layer.h"
 
 #include <cloud/filestore/libs/storage/model/public.h>
