+++ cloud/filestore/libs/storage/tablet/model/mixed_blocks.h
@@ -6,6 +6,7 @@
 #include "block.h"
 #include "block_list.h"
 #include "deletion_markers.h"
+#include "layer.h"
 
 #include <util/generic/vector.h>
 #include <util/memory/alloc.h>
