+++ cloud/filestore/libs/storage/tablet/model/ya.make
@@ -25,6 +25,7 @@ SRCS(
     group_by.cpp
     internal_request_id.cpp
     large_blocks.cpp
+    layer.cpp
     metadata_cache.cpp
     mixed_blocks.cpp
     node_ref.cpp
